Run ESMF_RegridWeightGen on single processor
Run E3SM_RegridWeightGen on single processor instead of using srun
and multiple nodes. srun was causing cryptic errors, usually claiming to
be out-of-memory errors, but the 3-30km mesh creation worked fine on
one processor. 1-10km case has not yet been tested, and may require the
issue with srun to be resolved.
Also do not create new scrip files for datasets if they already exist in
the data directory.